+++ packages/atomic/src/components/atomic-no-results/atomic-no-results.tsx
@@ -67,7 +67,7 @@ export class AtomicNoResults {
   private wrapHighlight(content: string) {
     return `<span class="font-bold" part="highlight">${this.bindings.i18n.t(
       'between-quotations',
-      {text: content}
+      {text: escape(content), interpolation: {escapeValue: false}}
     )}</span>`;
   }
 
@@ -93,7 +93,7 @@ export class AtomicNoResults {
     const content = this.querySummaryState.hasQuery
       ? this.bindings.i18n.t('no-results-for', {
           interpolation: {escapeValue: false},
-          query: this.wrapHighlight(escape(this.querySummaryState.query)),
+          query: this.wrapHighlight(this.querySummaryState.query),
         })
       : this.bindings.i18n.t('no-results');
     return (