+++ packages/atomic/src/components/atomic-smart-snippet/atomic-smart-snippet.tsx
@@ -12,10 +12,8 @@ import {
 } from '@coveo/headless';
 import {Hidden} from '../common/hidden';
 import {Heading} from '../common/heading';
-import {LinkWithResultAnalytics} from '../result-link/result-link';
 import {SmartSnippetFeedbackBanner} from './atomic-smart-snippet-feedback-banner';
 import {randomID} from '../../utils/utils';
-
 /**
  * The `atomic-smart-snippet` component displays the excerpt of a document that would be most likely to answer a particular query.
  *
@@ -121,6 +119,11 @@ export class AtomicSmartSnippet implements InitializableComponent {
     return styleTag.innerHTML;
   }
 
+  private set hideDuringRender(shouldHide: boolean) {
+    this.host.style.visibility = shouldHide ? 'hidden' : '';
+    this.host.style.position = shouldHide ? 'absolute' : '';
+  }
+
   private renderQuestion() {
     return (
       <Heading
@@ -138,53 +141,14 @@ export class AtomicSmartSnippet implements InitializableComponent {
       <atomic-smart-snippet-expandable-answer
         exportparts="answer,show-more-button,show-less-button,truncated-answer"
         part="body"
+        htmlContent={this.smartSnippetState.answer}
+        expanded={this.smartSnippetState.expanded}
         maximumHeight={this.maximumHeight}
         collapsedHeight={this.collapsedHeight}
         snippetStyle={this.style}
+        onExpand={() => this.smartSnippet.expand()}
+        onCollapse={() => this.smartSnippet.collapse()}
       ></atomic-smart-snippet-expandable-answer>
-    );
-  }
-
-  private renderSource() {
-    const {source} = this.smartSnippetState;
-    if (!source) {
-      return [];
-    }
-    return (
-      <section aria-label={this.bindings.i18n.t('smart-snippet-source')}>
-        <div part="source-url">
-          <LinkWithResultAnalytics
-            title={source.clickUri}
-            href={source.clickUri}
-            target="_self"
-            onSelect={() => this.smartSnippet.selectSource()}
-            onBeginDelayedSelect={() =>
-              this.smartSnippet.beginDelayedSelectSource()
-            }
-            onCancelPendingSelect={() =>
-              this.smartSnippet.cancelPendingSelectSource()
-            }
-          >
-            {source.clickUri}
-          </LinkWithResultAnalytics>
-        </div>
-        <div part="source-title" class="mb-6">
-          <LinkWithResultAnalytics
-            title={source.title}
-            href={source.clickUri}
-            target="_self"
-            onSelect={() => this.smartSnippet.selectSource()}
-            onBeginDelayedSelect={() =>
-              this.smartSnippet.beginDelayedSelectSource()
-            }
-            onCancelPendingSelect={() =>
-              this.smartSnippet.cancelPendingSelectSource()
-            }
-          >
-            {source.title}
-          </LinkWithResultAnalytics>
-        </div>
-      </section>
     );
   }
 
@@ -210,9 +174,17 @@ export class AtomicSmartSnippet implements InitializableComponent {
     }
   }
 
+  public componentDidRender() {
+    this.hideDuringRender = false;
+  }
+
   public render() {
     if (!this.smartSnippetState.answerFound) {
       return <Hidden></Hidden>;
+    }
+
+    if (this.host.classList.contains('atomic-hidden')) {
+      this.hideDuringRender = true;
     }
 
     return (
@@ -227,7 +199,18 @@ export class AtomicSmartSnippet implements InitializableComponent {
           {this.renderQuestion()}
           {this.renderContent()}
           <footer part="footer">
-            {this.renderSource()}
+            {this.smartSnippetState.source && (
+              <atomic-smart-snippet-source
+                source={this.smartSnippetState.source}
+                onSelectSource={() => this.smartSnippet.selectSource()}
+                onBeginDelayedSelectSource={() =>
+                  this.smartSnippet.beginDelayedSelectSource()
+                }
+                onCancelPendingSelectSource={() =>
+                  this.smartSnippet.cancelPendingSelectSource()
+                }
+              ></atomic-smart-snippet-source>
+            )}
             {this.renderFeedbackBanner()}
           </footer>
         </article>