+++ packages/atomic/src/components/facets/atomic-color-facet/atomic-color-facet.tsx
@@ -150,7 +150,7 @@ export class AtomicColorFacet
   // @Prop() public customSort?: string; TODO: KIT-753 Add customSort option for facet
 
   /**
-   * The required facets & values for this facet to be displayed.
+   * The required facets and values for this facet to be displayed.
    * Examples:
    * ```html
    * <atomic-facet facet-id="abc" field="objecttype" ...></atomic-facet>
@@ -168,7 +168,7 @@ export class AtomicColorFacet
    * ></atomic-color-facet>
    * ```
    */
-  @MapProp() public dependsOn: Record<string, string> = {};
+  @MapProp() @Prop() public dependsOn: Record<string, string> = {};
 
   @FocusTarget()
   private showMoreFocus!: FocusTargetController;