+++ packages/atomic/src/components/facets/atomic-rating-facet/atomic-rating-facet.tsx
@@ -143,7 +143,7 @@ export class AtomicRatingFacet
   @Prop({reflect: true}) public injectionDepth = 1000;
 
   /**
-   * The required facets & values for this facet to be displayed.
+   * The required facets and values for this facet to be displayed.
    * Examples:
    * ```html
    * <atomic-facet facet-id="abc" field="objecttype" ...></atomic-facet>
@@ -161,7 +161,7 @@ export class AtomicRatingFacet
    * ></atomic-rating-facet>
    * ```
    */
-  @MapProp() public dependsOn: Record<string, string> = {};
+  @MapProp() @Prop() public dependsOn: Record<string, string> = {};
 
   @FocusTarget()
   private headerFocus!: FocusTargetController;