+++ packages/quantic/cypress/integration/refine-modal-content/refine-modal-content-expectations.ts
@@ -35,6 +35,8 @@ const TIMEFRAME_FACET_PROPERTIES = [
   'injectionDepth',
 ];
 
+const DATE_FACET_PROPERTIES = ['numberOfValues', 'formattingFunction'];
+
 function refineContentExpectations(selector: RefineContentSelector) {
   return {
     displayFiltersTitle: () => {
@@ -139,6 +141,50 @@ function refineContentExpectations(selector: RefineContentSelector) {
           });
         })
         .logDetail('should display the duplicated timeframe facet');
+    },
+
+    displayDuplicatedTimeframeFacetValues: () => {
+      selector
+        .timeframeFacetValues()
+        .should('exist')
+        .then((elements) => {
+          const duplicatedFacetValues = Cypress.$.makeArray(elements).map(
+            (duplicatedElement) => duplicatedElement.innerText
+          );
+          selector.container
+            .timeframeFacetValues()
+            .should('exist')
+            .then((elements) => {
+              const facetValues = Cypress.$.makeArray(elements).map(
+                (element) => element.innerText
+              );
+              expect(duplicatedFacetValues).to.eql(facetValues);
+            });
+        })
+        .logDetail('should display the duplicated timeframe facet values');
+    },
+
+    displayDuplicatedDateFacet: () => {
+      selector
+        .dateFacet()
+        .should('exist')
+        .then((duplicatedFacet) => {
+          selector.container.dateFacet().then((facet) => {
+            [...COMMON_FACET_PROPERTIES, ...DATE_FACET_PROPERTIES].forEach(
+              (property) => {
+                expect(duplicatedFacet[0]).to.have.property(
+                  property,
+                  facet[0][property]
+                );
+              }
+            );
+            expect(duplicatedFacet[0].formattingFunction).to.be.eq(
+              facet[0].formattingFunction
+            );
+            expect(duplicatedFacet[0]).to.have.property('isCollapsed', true);
+          });
+        })
+        .logDetail('should display the duplicated date facet');
     },
 
     displaySort: (display: boolean) => {