+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,77 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.56.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.55.1...@coveo/atomic@1.56.0) (2022-05-12)
+
+
+### Features
+
+* **atomic:** Highlights smart snippet and smart snippet suggestion titles ([#1962](https://github.com/coveo/ui-kit/issues/1962)) ([41306f9](https://github.com/coveo/ui-kit/commit/41306f9e3663033c8c3d0be94e4b68555465aa1c))
+
+
+
+
+
+## [1.55.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.55.0...@coveo/atomic@1.55.1) (2022-05-11)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.55.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.54.0...@coveo/atomic@1.55.0) (2022-05-11)
+
+
+### Features
+
+* **atomic:** network improvements ([#1981](https://github.com/coveo/ui-kit/issues/1981)) ([6490e0d](https://github.com/coveo/ui-kit/commit/6490e0dd0ca16e28e3aa715b87ed2cdeb66047f1))
+
+
+
+
+
+# [1.54.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.53.1...@coveo/atomic@1.54.0) (2022-05-11)
+
+
+### Bug Fixes
+
+* **atomic:** added map props to atomic-react and atomic-angular ([#1976](https://github.com/coveo/ui-kit/issues/1976)) ([661faf1](https://github.com/coveo/ui-kit/commit/661faf1ec476d99feec95944a75dacfdcf38038a))
+
+
+### Features
+
+* **atomic,headless:** add more out of the box analytics configuration for atomic ([#1977](https://github.com/coveo/ui-kit/issues/1977)) ([7dff239](https://github.com/coveo/ui-kit/commit/7dff239d942140ef2f085e7aecbf21f1db039c8c))
+
+
+
+
+
+## [1.53.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.53.0...@coveo/atomic@1.53.1) (2022-05-10)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.53.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.52.0...@coveo/atomic@1.53.0) (2022-05-10)
+
+
+### Bug Fixes
+
+* **atomic:** made smart snippet wait for answer to be rendered ([#1971](https://github.com/coveo/ui-kit/issues/1971)) ([02fdf35](https://github.com/coveo/ui-kit/commit/02fdf35692f43d85e58544457e7074c8b66f3d11))
+* **atomic:** send correct result to rendering func and assign it at right time ([#1972](https://github.com/coveo/ui-kit/issues/1972)) ([857d62d](https://github.com/coveo/ui-kit/commit/857d62d8ce59ee5a4a0806aa4832f8682fb9157e))
+
+
+### Features
+
+* **atomic:** improved ripple animation speed ([#1974](https://github.com/coveo/ui-kit/issues/1974)) ([4623e96](https://github.com/coveo/ui-kit/commit/4623e961c73a706894a534606a7a366a30c46bec))
+
+
+
+
+
 # [1.52.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.51.0...@coveo/atomic@1.52.0) (2022-05-09)